Tighten requested GitHub access
Previously this requested `repo` [scope] when authenticating with
GitHub. That provides *full* access to both public and private
repositories, but we only want read access.

This switches to _no scope_, which only allows read access to public
information.

Unfortunately, it does not seem to be possible to request read-only
access to metadata about public and private repositories as an OAuth
app. In the future, switching to GitHub app authentication may allow
more sensible access controls.

[scope]: https://docs.github.com/en/apps/oauth-apps/building-oauth-apps/scopes-for-oauth-apps#available-scopes