`generate-static.ts`: Generate a JSON file
Previously this tried to inject the JSON into the HTML file as a field
on `window`, but that didn’t work with TypeScript/deno. Injecting it
into the compiled JavaScript caused problems because the TypeScript
compiler saw the empty array (pre-insertion) and removed various checks
based on that.

In a future commit we will load the file into the static app.

This also cleans up the code Claude generated a bit.