Indicate loading with animated button
This removes the loading message and adss a progress and pulse animation
on the “Refresh” button. The button label changes to “Loading”.

Progress is calculcated by adding up the specific contributions in the
the GitHub query response and comparing it to the summary total. This
happens when it is being processed into the `Calendar` for efficiency.
The calculation ignores existing data in `Calendar`.

Fixes: #58 — Loading message moves page too much